+++ src/components/SearchBar/Links.vue
@@ -1,0 +1,105 @@
+<template>
+  <div class="links">
+    <div class="link" v-for="link in links" :key="link.id">
+      <div class="regular">{{ link.regularLink }}</div>
+      <div class="action">
+        <div class="shorter">{{ link.shorterLink }}</div>
+        <button
+          @click="btnStatus(link.id, link.shorterLink)"
+          class="btn btn-copy"
+          :class="link.isActive ? 'isActive' : ''"
+        >
+          {{ link.isActive ? "Copied" : "Copy" }}
+        </button>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script>
+import { defineComponent, isMounted, ref } from "vue";
+import { uuid, validateInput, parseLink } from "./helpers.js";
+export default defineComponent({
+  props: { links: Array },
+  setup(props) {
+    const links = ref("links");
+    const shorten = (inputValue) => {
+      //   links.push({
+      //     id: uuid(),
+      //     regularLink: validateInput(inputValue),
+      //     shorterLink: parseLink(inputValue),
+      //     isActive: false,
+      //   });
+    };
+    const btnStatus = (id, copied) => {
+      //   console.log("btnstatus");
+      //   let copiedLinks = [...links];
+      //   copiedLinks.find((listItem) => {
+      //     listItem.id === id
+      //       ? (listItem.isActive = true)
+      //       : (listItem.isActive = false);
+      //   });
+    };
+    isMounted(() => {
+      console.log(links);
+      console.log("GG");
+    });
+    return { shorten, btnStatus, links };
+  },
+});
+</script>
+
+<style lang="scss">
+.links {
+  margin: 15px 0;
+  font-size: 18px;
+  display: flex;
+  flex-direction: column;
+
+  & .action {
+    display: flex;
+    align-items: center;
+    & .shorter {
+      margin-right: 10px;
+    }
+    & .btn-copy {
+      padding: 10px 20px;
+    }
+  }
+  .link {
+    box-sizing: border-box;
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background: #fff;
+    margin-bottom: 15px;
+    padding: 10px 20px;
+    border-radius: 4px;
+    color: hsl(180, 66%, 49%);
+    & .btn {
+      font-weight: bold;
+    }
+    & .regular {
+      color: hsl(257, 27%, 26%) !important;
+    }
+  }
+}
+.btn,
+.input {
+  cursor: pointer;
+  border-radius: 8px;
+  margin: 0 8px;
+}
+.btn {
+  background: hsl(180, 66%, 49%);
+  color: #fff;
+  border: none;
+  &:hover {
+    background: #9be3e2;
+  }
+}
+.btn-action {
+  font-weight: bold;
+}
+</style>
